添加 快选金额类型为int[]

+++ mobile-api/java/so/wwb/gamebox/mobile/controller/BaseDepositController.java
@@ -453,8 +453,8 @@ public class BaseDepositController {
     }
 
     //公司入款快选金额
-    public double[] quickSelection() {
-        double[] quickMoneys = new double[]{100, 200, 500, 1000, 5000};
+    public int[] quickSelection() {
+        int[] quickMoneys = new int[]{100, 200, 500, 1000, 5000};
         return quickMoneys;
     }
